step 14 1-5

+++ directory.rb
@@ -2,10 +2,9 @@
 
 def try_load_students
   filename = ARGV.first
-  return if filename.nil?
+  filename = 'students.csv' if filename.nil?
   if File.exist?(filename)
-    load_students(filename)
-    puts "Loaded #{@students.count} from #{filename}"
+    load_students(filename, false)
   else
     puts "Sorry, #{filename} doesn't exsists"
     exit
@@ -20,40 +19,39 @@ def interactive_menu
 end
 
 def print_menu
-  puts '1. Input the students'
-  puts '2. Show the students'
-  puts '3. Save the list to students.csv'
-  puts '4. Load the list from students.csv'
-  puts '9. Exit'
+  puts ['1. Input the students',
+        '2. Show the students',
+        '3. Save the list to students.csv',
+        '4. Load the list from students.csv',
+        '9. Exit']
 end
 
 def process(selection)
   case selection
-  when '1'
-    input_students
-  when '2'
-    show_students
-  when '3'
-    save_students
-  when '4'
-    load_students
-  when '9'
-    exit
+  when '1' then input_students
+  when '2' then show_students
+  when '3' then save_students
+  when '4' then load_students
+  when '9' then exit
   else
     puts "I don't know what you meant, try again"
   end
 end
 
 def input_students
-  puts 'Please enter the names of the students'
-  puts 'To finish, just hit return twice'
+  puts ['Please enter the names of the students',
+        'To finish, just hit return twice']
   name = STDIN.gets.chomp
   until name.empty?
-    @students << { name: name, cohort: :november }
+    insert_students(name)
     puts "Now we have #{@students.count} students"
     name = STDIN.gets.chomp
   end
-  @students
+  input_success
+end
+
+def insert_students(name, cohort = 'november')
+  @students << { name: name, cohort: cohort.to_sym }
 end
 
 def show_students
@@ -78,22 +76,53 @@ def print_footer
 end
 
 def save_students
-  file = File.open('students.csv', 'w')
+  filename = input_filename('s')
+  file = File.open(filename, 'w')
   @students.each do |student|
     student_data = [student[:name], student[:cohort]]
     csv_line = student_data.join(',')
     file.puts csv_line
   end
   file.close
+  save_success(filename)
 end
 
-def load_students(filename = 'students.csv')
-  file = File.open(filename, 'r')
-  file.readlines.each do |line|
-    name, cohort = line.chomp.split(',')
-    @students << {name: name, cohort: cohort.to_sym}
+def load_students(filename = 'students.csv', input = true)
+  filename = input_filename('l') if input
+  if File.exist?(filename)
+    @students = [] # rewrites the original @students
+    file = File.open(filename, 'r')
+    file.readlines.each do |line|
+      name, cohort = line.chomp.split(',')
+      insert_students(name, cohort)
+    end
+    file.close
+    load_success(filename)
+  else
+    puts 'No such a file! Loading was unsuccessful.'
   end
-  file.close
+end
+
+def load_success(filename)
+  puts "Loaded #{@students.count} from #{filename}"
+end
+
+def save_success(filename)
+  puts "Saved #{@students.count} student(s) into #{filename}."
+end
+
+def input_success
+  puts 'Inputting was successful.'
+end
+
+def input_filename(mode)
+  str_data = ['Enter a filename.',
+              'If no filename is given, it will be saved into students.csv.',
+              'If no filename is given, it will be loaded from students.csv.']
+  puts [str_data[0],(mode == 's' ? str_data[1] : str_data[2])]
+  filename = gets.chomp
+  filename = 'students.csv' if filename.empty?
+  filename
 end
 
 try_load_students
